requirements.txt
bettercam
Pillow
opencv-python
numpy
torch
ultralytics
configparser
